Create a mission report with the INBO corporate identity

Create a mission report with the INBO corporate identity

inbo_zending(conference, conferencedate, conferenceplace, website = "",
  reportdate, colleagues = "", floatbarrier = c(NA, "section", "subsection",
  "subsubsection"), citation_package = c("natbib", "none"), includes = NULL,
  codesize = c("footnotesize", "scriptsize", "tiny", "small", "normalsize"),
  lang = "dutch", keep_tex = FALSE, fig_crop = TRUE, pandoc_args = NULL,
  ...)

Arguments

conference
the name of the conference or workshop

conferencedate
the date of the conference

conferenceplace
the location of the conference

website
the website of the conference

reportdate
the date of this report

colleagues
name of other colleagues attending the same conference

floatbarrier
Should float barriers be placed? Defaults to NA (no extra float barriers). Options are "section", "subsection" and "subsubsection".

citation_package
The LaTeX package to process citations, natbib or biblatex. Use none if neither package is to be used.

includes
Named list of additional content to include within the document (typically created using the includes function).

codesize
The fontsize of the code, relative to the fontsize of the text (= normal size). Allowed values are "normalsize", "small", "footnotesize", "scriptsize" and "tiny". Defaults to "footnotesize".

lang
The language of the document. Defaults to "dutch"

keep_tex
Keep the tex file. Defaults to FALSE.

fig_crop
TRUE to automatically apply the pdfcrop utility (if available) to pdf figures

pandoc_args
Additional command line options to pass to pandoc

...
extra parameters: see details

Details

Available extra parameters:

hyphenation: the correct hyphenation for certain words