Create a report with the Flemish corporate identity

Create a report with the Flemish corporate identity

Usage

pdf_report(
  fig_crop = "auto",
  includes = NULL,
  keep_tex = FALSE,
  clean_supporting = TRUE,
  pandoc_args = NULL,
  ...
)

Arguments

fig_crop

Whether to crop PDF figures with the command pdfcrop. This requires the tools pdfcrop and ghostscript to be installed. By default, fig_crop = TRUE if these two tools are available.

includes

Named list of additional content to include within the document (typically created using the includes function).

keep_tex

Keep the intermediate tex file used in the conversion to PDF. Note that this argument does not control whether to keep the auxiliary files (e.g., .aux) generated by LaTeX when compiling .tex to .pdf. To keep these files, you may set options(tinytex.clean = FALSE).

clean_supporting

Cleanup any supporting files after conversion see render_supporting_files

pandoc_args

Additional command line options to pass to pandoc

...

currently ignored

Available YAML headers options shared among output formats

title: the main title

subtitle: the optional subtitle

author: the authors formatted as indicated in the section below.

flandersfont: if set to TRUE, use the Flanders Art Sans font. When not set, or set to FALSE or empty, use the Calibri font.

style: Defaults to "INBO" when missing.

"INBO" for the INBO style.

"Vlaanderen" for the generic style of the Flemish Government in Dutch.

"Flanders" for the generic style of the Flemish Government in a language other than Dutch.

lang: the main language of the document as RFC 5646 tags. style "INBO" and "Vlaanderen" require "nl" as lang. style "Flanders" uses "en" as default and can use other languages (but not "nl").

"nl" for Dutch

"en" for English

"fr" for French

codesize: Relative size of R code compared to normal text. Defaults to "footnotesize". All options going from large to small are: "normalsize", "small", "footnotesize", "scriptsize", "tiny". "normalsize" implies the same size as normal text.

Formatting author information

Quotes are required when a part contains spaces. If the author doesn't have an ORCID, you can omit that line. However, note that the INBO policy is that all scientific personnel is required to obtain an ORCID from https://orcid.org. Add corresponding: true to the corresponding author.

  - name:
     given: "Authors first name"
     family: "Authors last name"
    email: "local-part@domain.com"
    orcid: 0000-0002-1825-0097
    corresponding: true

Available YAML headers options specific for reports

Bold items are required when creating a public report. Omitting them results in an ugly warning in the rendered document. This is deliberate done so you can prepare the document without adding the information. But the ugly warnings nudges you to add the required information before publication.

public_report: A logical value indicating whether the report will be published or not. Setting public_report: FALSE will alter the colophon. Omitting public_report is equivalent to public_report: TRUE. An internal report can't have a DOI. Setting a DOI is equivalent to public_report: TRUE.

reviewer: names of the reviewers. Use the same syntax as for the authors.

year: year of publication. Used in the same places as shortauthor.

cover_photo: the relative path to the image you want on the cover.

cover_description: Description of the cover photo, including information on the author of the picture and license information.

cover the relative path to a pdf with the cover provided by the graphical designer. The first page of this pdf will be prepended to the pdf version of the report. It becomes the title page in the web and e-book version of the report. And also shows as a thumbnail in the floating table of content in the web version.

doi: the Digital Object Identifier of the report. Used in conjunction with shortauthor and year.

reportnr: Report number assigned by the INBO library.

ordernr: Optional reference number specified by the client.

depotnr: Report number assigned by the INBO library.

client: Optionally the name, address and website of the client.

client_logo: Optionally the logo of the client. Only used when client is set.

cooperation: Optionally the name, address and website of the organisation you collaborated with.

cooperation_logo: Optionally the logo of the collaborator. Only used when cooperation is set.

geraardsbergen: Set this to TRUE to add the address of INBO Geraardsbergen to the colophon. If not set or set to FALSE add the address of INBO Brussels.

watermark: an optional text to display as a watermark on the pdf or html. Note that omitting any of the required element will automatically generate a watermark with the text "DRAFT". Fill all required fields to get ride of this automatic watermark.

Available YAML headers options specific for reports in pdf format

Bold items are required. Omitting them results in an ugly warning in the rendered document. This is deliberate done so you can prepare the document without adding the information. But the ugly warnings nudges you to add the required information before publication.

embargo: The date at which the report is made public on the INBO website.

print: section only required in case you need a number of printed copies. When set it requires the following sub items:

copies: the number of copies

motivation: motivate why you need printed copies as the INBO policy is to use only digital publications.

pages: the number of pages in the full document.

tocdepth: which level headers to display.

0: up to chapters (#)

1: up to section (##) default

2: up to subsection (###)

lof: Adds a list of figures to the pdf after the table of content when set to TRUE. Omits the list of figures when missing or set to FALSE.

lot: Adds a list of tables to the pdf after the table of content and the list of figures when set to TRUE. Omits the list of tables when missing or set to FALSE.

hyphenation: An optional list of words with a specified hyphenation pattern. E.g. hyphenation: "fortran, ergo-no-mic" forces the word "fortran" to remain as a single word. "ergonomic" can be either the entire word, split into "ergo" and "nomic" or split into "ergono" and "mic".

floatbarrier: A float barrier forces to place all floating figures and tables before placing the remainder of the document. You can place them manually by setting \floatbarrier. This option allows you to set the automatically before a heading. Defaults to NA (only float barriers before starting a new chapter #). Options are "section" (before # and ##), "subsection" (before #, ## and ###) and "subsubsection" (before #, ##, ### and ####).

other_lang: the other languages you want to use in the pdf version. Available options are "nl", "en" and "fr" Defaults to all available languages except the main language.

Multiple languages in pdf

You can define some parts of the text to be in a different language than the main language (e.g. an abstract in a different language). This is currently limited to Dutch, English and French. Use \bdutch before and \edutch after the text in Dutch. Use \benglish before and \eenglish after the text in English. Use \bfrench before and \efrench after the text in French. The styles "INBO" and "Vlaanderen" have French and English as optional languages. The Flanders style has by default Dutch and French as optional languages.

Setting the language affects the hyphenation pattern and the names of items like figures, tables, table of contents, list of figures, list of tables, references, page numbers, ...

See also