Deprecated functions

old function	new function

ìnbo_ebook()	epub_book()

ìnbo_handouts()	handouts()

ìnbo_poster()	poster()

ìnbo_rapport()	pdf_report()

inbo_rapport_css()	report_template()

ìnbo_slides()	slides()

ìnbo_verslag()	minutes()

report()	pdf_report()

kolom_breedte(aantal = 3, marge = 20, witruimte = 10)

kolom_start(aantal = 3, start = 1, marge = 20, witruimte = 10)

ebook()

inbo_ebook()

inbo_handouts(
  toc = TRUE,
  nup = c("8", "1", "1plus", "2", "3", "3plus", "4", "4plus", "6"),
  ...
)

inbo_rapport_css(format = "html")

inbo_verslag(
  present = "",
  absent = "",
  chair = "",
  floatbarrier = c(NA, "section", "subsection", "subsubsection"),
  includes = NULL,
  codesize = c("footnotesize", "scriptsize", "tiny", "small", "normalsize"),
  lang = "dutch",
  keep_tex = FALSE,
  fig_crop = TRUE,
  pandoc_args = NULL,
  ...
)

inbo_zending(
  conference,
  conferencedate,
  conferenceplace,
  website = "",
  reportdate,
  colleagues = "",
  floatbarrier = c(NA, "section", "subsection", "subsubsection"),
  includes = NULL,
  codesize = c("footnotesize", "scriptsize", "tiny", "small", "normalsize"),
  lang = "dutch",
  keep_tex = FALSE,
  fig_crop = TRUE,
  pandoc_args = NULL,
  ...
)

inbo_rapport(fig_crop = "auto", includes = NULL, pandoc_args = NULL, ...)

report(fig_crop = "auto", includes = NULL, pandoc_args = NULL, ...)

inbo_poster(
  subtitle,
  codesize = c("footnotesize", "scriptsize", "tiny", "small", "normalsize"),
  lang = "english",
  email = "info@inbo.be",
  keep_tex = FALSE,
  fig_crop = TRUE,
  includes = NULL,
  pandoc_args = NULL,
  ...
)

inbo_slides(toc = TRUE, ...)

Arguments

aantal

The desired number of columns with equal width.

marge

the horizontal distance between columns in mm.

witruimte

the width of the whitespace edge around the poster in mm.

start

the number of the column

toc

display a table of content after the title slide

nup

Number of slides on a page. Defaults to 8. The 'plus' layouts leave blank space for recipients to make handwritten notes next to each slide. Note that the 2, 3plus, 4 and 6 layouts are intended for slides in the 4:3 aspect ratio, while the 3, 4plus and 8 layouts are intended for widescreen slides like the 16:10 aspect ratio.

...

currently ignored

format

defunct

present

names of all persons who attended the meeting

absent

names of all persons who couldn't make it

chair

chair of the meeting

floatbarrier

where to place automatic float barriers.

includes

Named list of additional content to include within the document (typically created using the includes function).

codesize

relative font size for code

lang

main language

keep_tex

Keep the intermediate tex file used in the conversion to PDF. Note that this argument does not control whether to keep the auxiliary files (e.g., .aux) generated by LaTeX when compiling .tex to .pdf. To keep these files, you may set options(tinytex.clean = FALSE).

fig_crop

Whether to crop PDF figures with the command pdfcrop. This requires the tools pdfcrop and ghostscript to be installed. By default, fig_crop = TRUE if these two tools are available.

pandoc_args

Additional command line options to pass to pandoc

conference

the name of the conference or workshop

conferencedate

the date of the conference

conferenceplace

the location of the conference

website

the website of the conference

reportdate

the date of this report

colleagues

name of other colleagues attending the same conference

subtitle

An optional subtitle

email

The email address to display at the bottom. Defaults to "info@inbo.be"