Create a minutes of a meeting with the INBO corporate identity

Create a minutes of a meeting with the INBO corporate identity

minutes(
  present = "",
  absent = "",
  chair = "",
  floatbarrier = c(NA, "section", "subsection", "subsubsection"),
  includes = NULL,
  codesize = c("footnotesize", "scriptsize", "tiny", "small", "normalsize"),
  lang = "dutch",
  keep_tex = FALSE,
  fig_crop = TRUE,
  pandoc_args = NULL,
  ...
)

Arguments

present

names of all persons who attended the meeting

absent

names of all persons who couldn't make it

chair

chair of the meeting

floatbarrier

where to place automatic float barriers.

includes

Named list of additional content to include within the document (typically created using the includes function).

codesize

relative font size for code

lang

main language

keep_tex

Keep the intermediate tex file used in the conversion to PDF. Note that this argument does not control whether to keep the auxiliary files (e.g., .aux) generated by LaTeX when compiling .tex to .pdf. To keep these files, you may set options(tinytex.clean = FALSE).

fig_crop

Whether to crop PDF figures with the command pdfcrop. This requires the tools pdfcrop and ghostscript to be installed. By default, fig_crop = TRUE if these two tools are available.

pandoc_args

Additional command line options to pass to pandoc

...

extra parameters: see details

Details

Available extra parameters:

hyphenation: the correct hyphenation for certain words

See also