Create a poster with the INBO theme version 2015

Create a poster with the INBO theme version 2015

poster(
  subtitle,
  codesize = c("footnotesize", "scriptsize", "tiny", "small", "normalsize"),
  lang = "english",
  email = "info@inbo.be",
  keep_tex = FALSE,
  fig_crop = TRUE,
  includes = NULL,
  pandoc_args = NULL,
  ...
)

Arguments

subtitle

An optional subtitle

codesize

relative font size for code

lang

The language of the document. Defaults to "english".

email

The email address to display at the bottom. Defaults to "info@inbo.be"

keep_tex

Keep the intermediate tex file used in the conversion to PDF. Note that this argument does not control whether to keep the auxiliary files (e.g., .aux) generated by LaTeX when compiling .tex to .pdf. To keep these files, you may set options(tinytex.clean = FALSE).

fig_crop

TRUE to automatically apply the pdfcrop utility (if available) to pdf figures

includes

Named list of additional content to include within the document (typically created using the includes function).

pandoc_args

Additional command line options to pass to pandoc.

...

extra parameters: see details

Details

Available extra parameters:

hyphenation: the correct hyphenation for certain words

flandersfont: Use the Flanders Art Sans font. Defaults to FALSE. Note that this requires the font to be present on the system.

ORCID: a list of authors. For each author there must a name and an ID. The ID is the author's ORCID ID , see https://orcid.org. This information will be displayed with QR code at the bottom of the poster.

DOI: a list of documents. For each document there must a name and an ID. The ID is the document's DOI, see https://doi.org. This information will be displayed with QR code at the bottom of the poster.

See also